Moving business name generation outside of loader.  Updated path.  Added logger
